Add option "spdy" enabling spdy patch
Option applies a patch from CloudFlare which re-enables SPDY in nginx
and allows it to co-exist with http2. This improves client compatability
and web performance for older pre-HTTP/2 clients which support SPDY.
This includes older versions of Chrome and Safari, and a number of
Android versions which are HTTP/1.1 and SPDY-only.